# documentation/api.yaml
openapi: 3.0.0
info:
  title: Documentation for the ERP @ablazelabs
  description: this documentation was made by swagger-ui
  version: "1.0"
  contact: 
    email: 'yaredterefeg@gmail.com'
  license:
    name: Ablaze Labs
    url: https://www.ablazelabs.com
paths:
  /account:
    post:
      summary: Creates a new user
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                email:
                  type: string
                phone_number:
                  type: string
                  description: countrycode-actual number(251-933221144)
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    get:
      summary: gets selected users
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                skip:
                  type: integer
                filter:
                  type: object
                  properties:
                    username:
                      type: string
                    email:
                      type: string
                    phone_number:
                      type: string
                    role:
                      type: number
                sort:
                  type: object
                  properties:
                    username:
                      type: integer
                    email:
                      type: integer
                    phone_number:
                      type: integer
                    id:
                      type: integer
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    id: 
                      type: integer
                    username:
                      type: string
                    normalized_username:
                      type: string
                    email:
                      type: string
                    phone_number:
                      type: string
                    two_factor_enabled:
                      type: boolean
                    lockout_enabled:
                      type: boolean
                    concurrency_stamp:
                      type: string
                    role: 
                      type: object
    patch:
      summary: updates a user
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                concurrency_stamp:
                  type: string
                id:
                  type: integer
                updateData:
                  type: object
                  properties:
                    username:
                      type: string
                    email:
                      type: string
                    phone_number:
                      type: string
                    role:
                      type: number
                    two_factor_enabled:
                      type: boolean
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    delete:
      summary: deletes a user
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /project/master/material:
    post:
      summary: Creates a new material
      tags: [Material]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                unit:
                  type: string
                material_category_id:
                  type: string
                startDate:
                  type: string
                endDate:
                  type: string
                isProtectedForEdit:
                      type: string

      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    get:
      summary: gets materials based on given filter and sort
      tags: [Material]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                skip:
                  type: integer
                filter:
                  type: object
                  properties:
                    name:
                      type: string
                    description:
                      type: string
                    unit:
                      type: string
                    material_category_name:
                      type: string
                sort:
                  type: object
                  properties:
                    name:
                      type: integer
                    description:
                      type: integer
                    unit:
                      type: integer
                    material_category_name:
                      type: integer
                    id:
                      type: integer
                    startDate:
                      type: integer
                    endDate:
                      type: integer
                    creationDate:
                      type: integer
                    revisionDate:
                      type: integer
                    isProtectedForEdit:
                      type: integer
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    description:
                      type: string
                    unit:
                      type: number
                    material_category:
                      type: object
                    startDate:
                      type: string
                    endDate:
                      type: string
                    creationDate:
                      type: string
                    createdBy:
                      type: string
                    revisionDate:
                      type: string
                    revisedBy:
                      type: string
                    isProtectedForEdit:
                      type: boolean                 
    patch:
      summary: updates a material
      tags: [Material]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                updateData:
                  type: object
                  properties:
                    name:
                      type: string
                    description:
                      type: string
                    unit:
                      type: string
                    material_category_id:
                      type: string
                    endDate:
                      type: string
                    startDate:
                      type: string
                    isProtectedForEdit:
                      type: string
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    delete:
      summary: deletes material
      tags: [Material]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...